Creating new layout to card view
