+++ package.json
@@ -53,6 +53,7 @@
     "material-ui": "^0.20.0",
     "mini-css-extract-plugin": "0.9.0",
     "nodemon": "^2.0.7",
+    "npm-watch": "^0.10.0",
     "optimize-css-assets-webpack-plugin": "5.0.3",
     "pnp-webpack-plugin": "1.6.4",
     "postcss-flexbugs-fixes": "4.1.0",
@@ -93,6 +94,7 @@
     "start": "node scripts/start.js",
     "express": "node scripts/express.js",
     "build": "node scripts/build.js",
+    "watch": "npm-watch",
     "test": "node scripts/test.js"
   },
   "eslintConfig": {
@@ -159,6 +161,14 @@
       "jest-watch-typeahead/filename",
       "jest-watch-typeahead/testname"
     ]
+  },
+  "watch": {
+    "build": {
+      "patterns": [
+        "src"
+      ],
+      "extensions": "js,jsx"
+    }
   },
   "babel": {
     "presets": [
